refactor(lsp): clean up completion resolve
Use the existing declaration position index for eager fallback and store structured resolved NatSpec instead of duplicate rendered strings. Keep protocol tests in a dedicated module while preserving stale and ambiguous item fallback.
