+++ crates/lsp/src/lib.rs
@@ -24,13 +24,13 @@ mod commands;
 mod config;
 mod diagnostics;
 mod document_links;
+mod documentation;
 mod file_operations;
 mod flycheck;
 mod folding_range;
 mod formatter;
 mod global_state;
 mod handlers;
-mod hover;
 mod inlay_hints;
 mod natspec_completion;
 mod override_index;
